user can guess all rows at once, remove individual row guessing
